requirements.txt
flask
line-bot-sdk
beautifulsoup4
requests
gspread
oauth2client
haversine